tests(Import): fix php 8.1 deprecation notice

See merge request tine20/tine20!4128

+++ tests/tine20/ImportTestCase.php
@@ -55,7 +55,7 @@ abstract class ImportTestCase extends TestCase
         }
         
         // cleanup
-        if (file_exists($this->_filename) && $this->_deleteImportFile) {
+        if ($this->_filename && file_exists($this->_filename) && $this->_deleteImportFile) {
              unlink($this->_filename);
         }
     }
